Job killing signal should also affect the stager https://github.com/DIRACGrid/DIRAC/issues/1510
Make sure to only cancel file staging for files with no relations with other tasks (jobs) but the killed ones!
"Canceling" is removing all information on these tasks, from the SMS DB.

+++ StorageManagementSystem/DB/StorageManagementDB.py
@@ -966,15 +966,16 @@ class StorageManagementDB( DB ):
          
       replicaIDs = [ row[0] for row in res['Value'] ]
       
-      req = "DELETE FROM StageRequests WHERE ReplicaID IN (%s);" % intListToString ( replicaIDs )      
-      res = self._update( req, connection )
-      if not res['OK']:
-        gLogger.error( "%s.%s_DB: problem removing records: %s. %s" % ( self._caller(), 'killTasksBySourceTaskID', req, res['Message'] ) )
+      if replicaIDs:
+        req = "DELETE FROM StageRequests WHERE ReplicaID IN (%s);" % intListToString ( replicaIDs )      
+        res = self._update( req, connection )
+        if not res['OK']:
+          gLogger.error( "%s.%s_DB: problem removing records: %s. %s" % ( self._caller(), 'killTasksBySourceTaskID', req, res['Message'] ) )
      
-      req = "DELETE FROM CacheReplicas WHERE ReplicaID in (%s) AND Links=1;" % intListToString ( replicaIDs )
-      res = self._update( req, connection )
-      if not res['OK']:
-        gLogger.error( "%s.%s_DB: problem removing records: %s. %s" % ( self._caller(), 'killTasksBySourceTaskID', req, res['Message'] ) )
+        req = "DELETE FROM CacheReplicas WHERE ReplicaID in (%s) AND Links=1;" % intListToString ( replicaIDs )
+        res = self._update( req, connection )
+        if not res['OK']:
+          gLogger.error( "%s.%s_DB: problem removing records: %s. %s" % ( self._caller(), 'killTasksBySourceTaskID', req, res['Message'] ) )
       
       # Finally, remove the Task and TaskReplicas entries.
       res = self.removeTasks(taskIDs, connection)
